feat: implement ext2 large file support (LARGE_FILE feature) with 64-bit inode sizes (Item 147)
- Add EXT2_FEATURE_RO_COMPAT_LARGE_FILE (0x0002) and related
  read-only compatible feature flag defines to ext2.h
- Add ext2_inode_get_size() helper that returns uint64_t file size
  by combining i_size with i_dir_acl (repurposed as i_size_high)
  when the LARGE_FILE feature is active and inode is a regular file
- Update all i_size references in ext2.c to use the helper:
  ext2_find_in_dir, ext2_read_dir, ext2_read, ext2_stat
- Change vfs_stat.size from uint32_t to uint64_t to accommodate
  large file sizes across all VFS filesystem implementations

+++ src/include/vfs.h
@@ -60,7 +60,7 @@ struct posix_acl {
 
 /* VFS stat result */
 struct vfs_stat {
-    uint32_t size;
+    uint64_t size;
     uint8_t  type;   /* 1=file, 2=dir */
     uint16_t uid;
     uint16_t gid;
